Refactor PkgFinishClient to inherit from Installation::FinishClient
* It now lives in lib/packager/pkg_finish.rb.
